app/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.10.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
